Fix test_minargmin.cpp example

+++ keops/sandbox/test_logsumexp.cpp
@@ -69,7 +69,7 @@ int main() {
 
     std::cout << "Testing LogSumExp reduction" << std::endl;
 
-    int Nx=5000, Ny=2000;
+    size_t Nx=5000, Ny=2000;
         
     std::vector<__TYPE__> vf(Nx*F::DIM);    fillrandom(vf); __TYPE__ *f = vf.data();
     std::vector<__TYPE__> vx(Nx*DIMPOINT);    fillrandom(vx); __TYPE__ *x = vx.data();
@@ -92,7 +92,7 @@ int main() {
 
     // display values
     std::cout << "rescpu1 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << rescpu1[i] << " ";
     std::cout << "..." << std::endl << std::endl;
     
@@ -102,7 +102,7 @@ int main() {
 
     begin = clock();
     Eval<SUMEXPF,CpuConv>::Run(Nx, Ny, f, oos2, x, y, b);
-    for(int i=0; i<vf.size(); i++)
+    for(size_t i=0; i<vf.size(); i++)
     	vf[i] = log(vf[i]);
     end = clock();
     std::cout << "time for CPU computation : " << double(end - begin) / CLOCKS_PER_SEC << std::endl;
@@ -112,13 +112,13 @@ int main() {
 
     // display values
     std::cout << "rescpu2 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << rescpu2[i] << " ";
     std::cout << "..." << std::endl << std::endl;
 
     // display mean of errors
     __TYPE__ s = 0;
-    for(int i=0; i<Nx*F::DIM; i++)
+    for(size_t i=0; i<Nx*F::DIM; i++)
         s += std::abs(rescpu1[i]-rescpu2[i]);
     std::cout << std::endl << "mean abs error = " << s/Nx << std::endl << std::endl;
 
@@ -143,7 +143,7 @@ int main() {
     rescpu1 = vf;
 
     std::cout << "rescpu1 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << rescpu1[i] << " ";
     std::cout << "..." << std::endl << std::endl;
 
@@ -151,19 +151,19 @@ int main() {
 	using GX2 = Grad<SUMEXPF,X,E>;
     begin = clock();
 	Eval<GX2,CpuConv>::Run(Nx, Ny, f, oos2, x, y, b, e);
-    for(int i=0; i<vf.size(); i++)
+    for(size_t i=0; i<vf.size(); i++)
     	vf[i] = vf[i]/exp(rescpu2[i/GX::DIM]);
     end = clock();
     std::cout << "time for CPU computation : " << double(end - begin) / CLOCKS_PER_SEC << std::endl;
 
     rescpu2 = vf;
     std::cout << "rescpu2 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << rescpu2[i] << " ";
     std::cout << "..." << std::endl << std::endl;
 
     // display mean of errors
-    for(int i=0; i<Nx*GX::DIM; i++)
+    for(size_t i=0; i<Nx*GX::DIM; i++)
         s += std::abs(rescpu1[i]-rescpu2[i]);
     std::cout << std::endl << "mean abs error = " << s/Nx << std::endl << std::endl;
